--- LedgerWallet/WindowsDevice.cs
@@ -1,12 +1,0 @@
-﻿using Hid.Net;
-using LedgerWallet.Transports;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace LedgerWallet
-{
-	public static class WindowsDevice
-	{
-
-	}
-}
